chore: build dietary-record document, move body related files and MealForm from modules.tsx
