+++ src/common/components/dietary-record/BodyRate.tsx
